# internal/transport/http/httpgen/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package httpgen

type APIV1MarketParserParseGetBadRequest ErrorResponse

func (*APIV1MarketParserParseGetBadRequest) aPIV1MarketParserParseGetRes() {}

type APIV1MarketParserParseGetCode499 ErrorResponse

func (*APIV1MarketParserParseGetCode499) aPIV1MarketParserParseGetRes() {}

type APIV1MarketParserParseGetGatewayTimeout ErrorResponse

func (*APIV1MarketParserParseGetGatewayTimeout) aPIV1MarketParserParseGetRes() {}

type APIV1MarketParserParseGetInternalServerError ErrorResponse

func (*APIV1MarketParserParseGetInternalServerError) aPIV1MarketParserParseGetRes() {}

// Ref: #/components/schemas/ErrorResponse
type ErrorResponse struct {
	Status  int    `json:"status"`
	Message string `json:"message"`
}

// GetStatus returns the value of Status.
func (s *ErrorResponse) GetStatus() int {
	return s.Status
}

// GetMessage returns the value of Message.
func (s *ErrorResponse) GetMessage() string {
	return s.Message
}

// SetStatus sets the value of Status.
func (s *ErrorResponse) SetStatus(val int) {
	s.Status = val
}

// SetMessage sets the value of Message.
func (s *ErrorResponse) SetMessage(val string) {
	s.Message = val
}

type ParseResponse []Product

func (*ParseResponse) aPIV1MarketParserParseGetRes() {}

// Ref: #/components/schemas/Product
type Product struct {
	Name  string  `json:"name"`
	Link  string  `json:"link"`
	Price float64 `json:"price"`
}

// GetName returns the value of Name.
func (s *Product) GetName() string {
	return s.Name
}

// GetLink returns the value of Link.
func (s *Product) GetLink() string {
	return s.Link
}

// GetPrice returns the value of Price.
func (s *Product) GetPrice() float64 {
	return s.Price
}

// SetName sets the value of Name.
func (s *Product) SetName(val string) {
	s.Name = val
}

// SetLink sets the value of Link.
func (s *Product) SetLink(val string) {
	s.Link = val
}

// SetPrice sets the value of Price.
func (s *Product) SetPrice(val float64) {
	s.Price = val
}